chore: moving card icon resolver in an util class for easier access
